allow the result of compression to be larger than the original
While unfortunate it is true that for some inputs LZO will produce a
larger output than the original.  This is especially likely to happen
for random bytes, or already compressed data, but in theory it can
happen for any other input.  In anycase some protocols and formats
require the data be LZO compressed even if uncompressed would be
smaller.  As such its necessary to return the compressed buffer even if
it is larger than the input.  Some consumers may choose to not use the
result if it is larger, but that choice belongs in the consumer, not a
wrapper around liblzo.